[GAL-213] Persist user changes on demand only
